#28 - Refactoring factories. Providing javadoc.
