+++ core-java/src/main/java/chessgame/domain/BoardColor.java
@@ -7,7 +7,14 @@ package chessgame.domain;
  * @since September 2016
  */
 public enum BoardColor {
-    LIGHT("#FFCE9E;"), DARK("#D18B47;");
+    /**
+     * Light color of a chess board position.
+     */
+    LIGHT("#FFCE9E;"),
+    /**
+     * Dark color of a chess board position.
+     */
+    DARK("#D18B47;");
 
     private final String color;
