+++ core-java/src/main/java/chessgame/domain/factory/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Factories of Chess game model.
+ */
+package chessgame.domain.factory;
