+++ core-java/src/main/java/chessgame/domain/model/BlackPawn.java
@@ -4,8 +4,21 @@ import chessgame.domain.Board;
 
 import static chessgame.domain.PieceColor.BLACK;
 
+/**
+ * Black pawn instances of a chess game board.
+ *
+ * @author Prestes, E. M.
+ * @see AbstractPawn
+ * @see Board
+ * @since August 2021
+ */
 public class BlackPawn extends AbstractPawn {
 
+    /**
+     * Constructor compose by a board.
+     *
+     * @param board Board informed
+     */
     public BlackPawn(Board board) {
         super(board, BLACK, "\u265F", "&#9823;");
     }
