--- core-java/src/main/java/chessgame/domain/model/BoardFactory.java
@@ -1,12 +1,0 @@
-package chessgame.domain.model;
-
-public final class BoardFactory {
-
-    private BoardFactory() {
-        super();
-    }
-
-    public static DefaultBoard createBoard() {
-        return new DefaultBoard();
-    }
-}
