+++ core-java/src/main/java/chessgame/domain/package-info.java
@@ -1,4 +1,4 @@
 /**
- * Interfaces and factories of Chess game model.
+ * Interfaces of Chess game model.
  */
 package chessgame.domain;
