+++ core-java/src/main/java/module-info.java
@@ -1,5 +1,8 @@
+/**
+ * Chess game core module.
+ */
 module chessgame.core {
 
     exports chessgame.domain;
-    exports chessgame.domain.model;
+    exports chessgame.domain.factory;
 }
