+++ core-java/src/test/java/chessgame/domain/model/package-info.java
@@ -1,0 +1,1 @@
+package chessgame.domain.model;
